+++ bowfill.go
@@ -8,9 +8,11 @@ import (
 	"github.com/apache/arrow/go/v8/arrow/array"
 )
 
+// TODO: add support for timestamp types
+
 // FillLinear fills the column toFillColIndex using the Linear interpolation method according
 // to the reference column refColIndex, which has to be sorted.
-// Fills only int64 and float64 types.
+// Fills only Int64 and Float64 types.
 func (b *bow) FillLinear(refColIndex, toFillColIndex int) (Bow, error) {
 	if refColIndex < 0 || refColIndex > b.NumCols()-1 {
 		return nil, fmt.Errorf("bow.FillLinear: refColIndex is out of range")