+++ bowparquet_test.go
@@ -16,40 +16,55 @@ const (
 
 func TestParquet(t *testing.T) {
 	t.Run("read/write input file", func(t *testing.T) {
-		bBefore, err := NewBowFromParquet(testInputFileName, true)
+		bBefore, err := NewBowFromParquet(testInputFileName, false)
 		assert.NoError(t, err)
-		fmt.Printf("bBefore\n%s\n", bBefore)
 
-		assert.NoError(t, bBefore.WriteParquet(testOutputFileName, true))
+		assert.NoError(t, bBefore.WriteParquet(testOutputFileName, false))
 
-		bAfter, err := NewBowFromParquet(testOutputFileName+".parquet", true)
+		bAfter, err := NewBowFromParquet(testOutputFileName+".parquet", false)
 		assert.NoError(t, err)
-		fmt.Printf("bAfter\n%s\n", bAfter)
 
 		assert.Equal(t, bBefore.String(), bAfter.String())
 
 		require.NoError(t, os.Remove(testOutputFileName+".parquet"))
 	})
 
-	t.Run("bow supported types with rows and nils", func(t *testing.T) {
+	t.Run("all supported types with rows and nil values", func(t *testing.T) {
 		bBefore, err := NewBowFromRowBasedInterfaces(
-			[]string{"int", "float", "bool", "string"},
-			[]Type{Int64, Float64, Bool, String},
+			[]string{"int", "float", "bool", "string",
+				"timestamp_ms_int", "timestamp_ms_str",
+				"timestamp_us_int", "timestamp_us_str",
+				"timestamp_ns_int", "timestamp_ns_str"},
+			[]Type{Int64, Float64, Bool, String,
+				TimestampMilli, TimestampMilli,
+				TimestampMicro, TimestampMicro,
+				TimestampNano, TimestampNano},
 			[][]interface{}{
-				{1, 1., true, "hi"},
-				{2, 2., false, "ho"},
-				{nil, nil, nil, nil},
-				{3, 3., true, "hu"},
+				{1, 1., true, "hi",
+					1651017600000, "2022-04-27T00:00:00.123Z",
+					1651017600000000, "2022-04-27T00:00:00.123456Z",
+					1651017600000000000, "2022-04-27T00:00:00.123456789Z"},
+				{2, 2., false, "ho",
+					1651021200000, "2022-04-27T01:00:00.123Z",
+					1651021200000000, "2022-04-27T01:00:00.123456Z",
+					1651021200000000000, "2022-04-27T01:00:00.123456789Z"},
+				{nil, nil, nil, nil, nil, nil, nil, nil, nil, nil},
+				{3, 3., true, "hu",
+					1651028400000, "2022-04-27T03:00:00.123Z",
+					1651028400000000, "2022-04-27T03:00:00.123Z456",
+					1651028400000000000, "2022-04-27T03:00:00.123456789Z"},
 			})
 		require.NoError(t, err)
+
+		fmt.Printf("bBefore\n%s\n", bBefore)
 
 		assert.NoError(t, bBefore.WriteParquet(testOutputFileName+"_withrows", true))
 
 		bAfter, err := NewBowFromParquet(testOutputFileName+"_withrows.parquet", true)
 		assert.NoError(t, err)
 
-		fmt.Printf("bBefore\n%s\n", bBefore)
 		fmt.Printf("bAfter\n%s\n", bAfter)
+
 		assert.Equal(t, bBefore.String(), bAfter.String())
 
 		require.NoError(t, os.Remove(testOutputFileName+"_withrows.parquet"))
@@ -62,9 +77,9 @@ func TestParquet(t *testing.T) {
 			[][]interface{}{})
 		require.NoError(t, err)
 
-		assert.NoError(t, bBefore.WriteParquet(testOutputFileName+"_norows", true))
+		assert.NoError(t, bBefore.WriteParquet(testOutputFileName+"_norows", false))
 
-		bAfter, err := NewBowFromParquet(testOutputFileName+"_norows.parquet", true)
+		bAfter, err := NewBowFromParquet(testOutputFileName+"_norows.parquet", false)
 		assert.NoError(t, err)
 
 		assert.Equal(t, bBefore.String(), bAfter.String())
@@ -76,7 +91,7 @@ func TestParquet(t *testing.T) {
 		bBefore := NewBowEmpty()
 
 		assert.Errorf(t,
-			bBefore.WriteParquet(testOutputFileName+"_empty", true),
+			bBefore.WriteParquet(testOutputFileName+"_empty", false),
 			"bow.WriteParquet: no columns",
 		)
 	})