+++ bowseries_test.go
@@ -14,7 +14,6 @@ func TestNewSeriesFromInterfaces(t *testing.T) {
 			testcase := []interface{}{typ.Convert(0), nil}
 			res, err := NewBow(NewSeriesFromInterfaces(typ.String(), typ, testcase))
 			require.NoError(t, err)
-			fmt.Printf("BOW\n%s\n", res)
 			assert.Equal(t, typ.Convert(0), res.GetValue(0, 0))
 			assert.Equal(t, nil, res.GetValue(0, 1))
 		})